requirement.txt
#absl-py==0.15.0
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
colorama==0.4.4
comtypes==1.1.10
cycler==0.11.0
decorator==5.1.0
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.2
google-pasta==0.2.0
graphviz==0.19
grpcio==1.34.1
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.1
imutils==0.5.4
ipython==7.29.0
jedi==0.18.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libopencv==0.0.1
libpng-bins==0.0.3
#libtiff==0.4.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
#mkl-fft==1.2.0
mkl-random==1.2.2
#numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
openpyxl==3.0.9
openssl-python==0.1.1
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.5
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
pywin32==302
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
